elasticsearch

+++ src/main/java/tr/com/huseyinari/elasticsearchdemo/dto/EmployeeDto.java
@@ -2,6 +2,7 @@ package tr.com.huseyinari.elasticsearchdemo.dto;
 
 import lombok.Getter;
 import lombok.Setter;
+import org.springframework.format.annotation.DateTimeFormat;
 
 import java.time.LocalDate;
 
@@ -10,7 +11,7 @@ import java.time.LocalDate;
 public class EmployeeDto {
     private String id;
     private String name;
-    private Long salary;
+    private Float salary;
     private Integer age;
     private LocalDate startDate;
     private boolean active;
